chatbot/requirement.txt
Flask
python-dotenv
langchain
langchain-community
openai==1.90.0
gunicorn
